[refactor] rename Common/Values.js to CommonVals.js and update imports
