+++ Framework/CommonVals.js
